UserId state using Redux, notes collection updated(particular user)

+++ src/components/noteCard/NoteCardTrash.jsx
@@ -2,16 +2,19 @@ import React from 'react';
 import "./NoteCardTrash.scss";
 import { FaEllipsisV, FaBookmark } from 'react-icons/fa';
 import FirestoreService from '../../services/firestoreService';
+import {useSelector} from "react-redux";
 
 export const NoteCardTrash = ({note}) => {
+    const userId = useSelector(state=>state.userId);
+
     return (
         <div id="Note-Card-Trash" onClick={()=>{}}>
             <h3 id="Note-Title">{note?.title} </h3>
             <div className="Dropdown">
                 < FaEllipsisV id="Dropdown-Icon"  size="18"/>
                 <div id="Dropdown-Options">
-                    <button onClick={()=>FirestoreService.trashNote({noteId: note?.id, isDeleted : false})} className="Dropdown-Btn">Restore</button>
-                    <button onClick={()=>FirestoreService.deleteNote({noteId: note?.id})} className="Dropdown-Btn">Delete Permanently</button>
+                    <button onClick={()=>FirestoreService.trashNote({noteId: note?.id, isDeleted : false, userId : userId})} className="Dropdown-Btn">Restore</button>
+                    <button onClick={()=>FirestoreService.deleteNote({noteId: note?.id, userId : userId})} className="Dropdown-Btn">Delete Permanently</button>
                 </div>
             </div>
             <p id="Note-Description">
